test: update imports in stories to new mock location
